Refactor NumOrder validators testing
